+++ slip.c
@@ -247,7 +247,7 @@ int slip_decode(struct slip_decode_state *state)
 	    dump("de-slipped packet",state->dst,state->packet_length);
 	 
 	  // Check that CRC matches
-	  unsigned long crc=Crc32_ComputeBuf( 0, state->dst, state->packet_length);
+	  uint32_t crc=Crc32_ComputeBuf( 0, state->dst, state->packet_length);
 	  if (crc!=state->crc) {
 	    if (config.debug.packetradio)
 	      DEBUGF("Rejected packet of %d bytes due to CRC mis-match (%08x vs %08x)",
